# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "BuyerAnalyticsResource" : {
        "properties" : {
          "companyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "completedPayments" : {
            "format" : "int64",
            "type" : "integer"
          },
          "monthlySpending" : {
            "items" : {
              "$ref" : "#/components/schemas/MonthlyAmount"
            },
            "type" : "array"
          },
          "pendingPayments" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalSpent" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "MonthlyAmount" : {
        "properties" : {
          "amount" : {
            "format" : "double",
            "type" : "number"
          },
          "month" : {
            "type" : "string"
          },
          "monthIndex" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PlatformSummaryResource" : {
        "properties" : {
          "completedDeliveries" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pendingOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalDeliveries" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPayments" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalRevenue" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "ProviderAnalyticsResource" : {
        "properties" : {
          "cancelledOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "confirmedOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "monthlyRevenue" : {
            "items" : {
              "$ref" : "#/components/schemas/MonthlyAmount"
            },
            "type" : "array"
          },
          "pendingOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "providerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "salesTrend" : {
            "items" : {
              "$ref" : "#/components/schemas/SalesTrendPoint"
            },
            "type" : "array"
          },
          "totalFuelSoldLitres" : {
            "format" : "double",
            "type" : "number"
          },
          "totalOrders" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalRevenue" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "SalesTrendPoint" : {
        "properties" : {
          "date" : {
            "format" : "date",
            "type" : "string"
          },
          "litres" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@fuelguard.app",
      "name" : "Soporte de FuelGuard",
      "url" : "https://fuelguard.app/support"
    },
    "description" : "API de FuelGuard para gestionar abastecimiento, flota, entregas, inventario, telemetría IoT y pagos por organización.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FuelGuard Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/analytics/buyers/{companyId}" : {
      "get" : {
        "description" : "Devuelve métricas agregadas de la empresa indicada cuando el usuario pertenece a ella.",
        "operationId" : "getBuyerAnalytics",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerAnalyticsResource"
                }
              }
            },
            "description" : "Indicadores de la empresa compradora devueltos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BuyerAnalyticsResource"
                }
              }
            },
            "description" : "El usuario no pertenece a la empresa compradora solicitada."
          }
        },
        "summary" : "Consultar indicadores de empresa compradora",
        "tags" : [ "Analítica" ]
      }
    },
    "/api/analytics/platform" : {
      "get" : {
        "description" : "Devuelve métricas agregadas de toda la plataforma; requiere autoridad ROLE_ADMIN.",
        "operationId" : "getPlatformSummary",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlatformSummaryResource"
                }
              }
            },
            "description" : "Resumen de plataforma devuelto."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlatformSummaryResource"
                }
              }
            },
            "description" : "El usuario no cuenta con autoridad ROLE_ADMIN."
          }
        },
        "summary" : "Consultar resumen de plataforma",
        "tags" : [ "Analítica" ]
      }
    },
    "/api/analytics/providers/{providerId}" : {
      "get" : {
        "description" : "ROLE_PROVIDER propietario. from/to ISO inclusivos: pedidos por creación UTC, ingresos por pago, litros vendidos una vez por orden con pago completado, en fecha de su primer pago. Sin fechas conserva los acumulados. salesTrend diario se puede agrupar semanal/mensualmente; unidades desconocidas se excluyen de litros.",
        "operationId" : "getProviderAnalytics",
        "parameters" : [ {
          "in" : "path",
          "name" : "providerId",
          "required" : true,
          "schema" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProviderAnalyticsResource"
                }
              }
            },
            "description" : "Indicadores del distribuidor devueltos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProviderAnalyticsResource"
                }
              }
            },
            "description" : "El usuario no pertenece al distribuidor solicitado."
          }
        },
        "summary" : "Consultar indicadores del distribuidor",
        "tags" : [ "Analítica" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}